requirements.txt
Django~=4.2.2
environs~=9.5.0
Pillow~=9.5.0
django-admin-sortable2~=2.1.8
requests~=2.31.0
pytils~=0.4.1
django-tinymce~=3.6.1
django-admin-sortable~=2.3
django-debug-toolbar==4.2.0